Refactor ZMQClient and its collaborators
Trying to get at a good implemtation and testability for ZMQClient. I
think this is better. The responsibilites are separated and the
dependencies can be injected for testing.

This hasn't yet been tested by having it communicate with an actual ZMQ
server, though. That set of integration tests is the next step.