Add integration test for making requests
The test in spec/integration/request_spec.rb creates a little ZMQ server
to respond to test requests. The end result is a test that checks the
intgration between the Mailman object that clients are using all the way
down to the lower-level ZMQ request/response handling.

All the test does is assert that the
response is not nil. It doesn't check that the response conforms to the
response interface.