Improving tests, response tests, response fix
Adding tests for responses, which required some reworking of the test
ZMQ Server as the thread would sometimes lead to failed tests or a hung
process. This current approach works, but it doesn't seem ideal.

Also changing the requester to expect a single-part response
`recv_string` instead of multi-part `recv_strings`. Though the response
parser should be able to handle responses like "200" or ["200"]. This
should be revisited once the actual repsonses have been determined.