Refactor to follow previous implementation pattern
First pass at Subscription ignored how we'd implemented the handling of
sending messages. This commit refactors Subscription to follow our
established approach.

RubyMailman::Subscription is now just a constructor/wrapper around the
CentralServicesInterface (which can be swapped out/ injected as needed).
Subscription is the outer edge of the system, so it does the parameter
verification.