Hide ZMQClient behind CSTransportClient
I didn't see a reason for CentralServiceInterface to know anything about
its transport client. We intend to use ZMQ for now, but it could change.

So I have CentralServiceInterface using a CSTransportClient, which I
configure at run-time to be the ZMQClient. Should we leave ZMQ, then the
CS files won't need to change, we'll just have to swap in the new
transport as CSTransportClient.