Add subscription, refactor ZMQClient, bust tests
This commit adds a working subscripton method. As specified in the
README. Current approach is to handle the receiving of updates in a
thread, though that could change.

Updates received on a subscribed channel are turned into Messages and
then passed to the listener.

Implementing the subscription exposed some bad design decisions in the
ZMQClient code. I think I prematurely abstracted stuff and it got weird
when dealing with a subscription. So I've blown all that stuff up,
re-introduced a lot of duplication and left it kind of ugly. Once tests
are working again, I can refactor that part of the gem into something
nicer.

That refactor broke most of the tests. And Subscription wasn't ever
really tested to begin with. So next step is to get the test suite back
in order.