Return RubyMailman::Subscription test to working
Getting these tests working again and tweaking how the class works. No
real need for a `subscribe` method, since it's always called. Might as
well put that in initialize and save ourselves a little stubbing.

Also explicitly requiring that the class-level `subscribe` returns an
instance of RubyMailman::Subscription. Right now there's nothing to do
with that instance, as it has no public methods. But this gives us a way
to add functionality later.