Test ZMQRequestClient
This test is a little odd. The previous implementation mocked everything
out and ensured that the ZMQ stuff was being set up correctly. But
that's all implementation and shouldn't be tested.

The only way to really test this is to have it talk to a real ZMQ server
that can handle REQ/REP. So this testing approach introduces a simple
REQ/REP server in `spec/zmq_servers/reply` that can be spun up before a
test and then killed afterwards.

This testing server can also be configured so that we can randomize
replies or change ports.

The resulting test in more integration-ish, since it talks to an
external service. But I'm keeping it in unit tests because that external
service is as stub-like as I can make it. The real integration tests
will test end-to-end: from the customer-facing API all the way down to a
ZMQServer instance.