Test ZMQSubscriptionClient
Following the example of testing the ZMQRequestClient. Again, this is an
integration-ish test because there has to be a real ZMQ Publisher
somewhere for the client to interact with.

The test publishing server is in `spec/zmq_servers/publish.rb` and is
fairly straight forward.

The test is a little goofier than the Request/Reply one because the
publisher is constantly spitting out messages, whereas REQ/REP is a 1:1
relationship. So the tests have a lot of "expect to be called at least
once" because the publisher might publish twice during the test's run.
And I didn't see a reason to fail the tests if that happened.

And I'm injecting the message_builder class for the same reasons I
outlined in an earlier commit. It allows us access to RSpec's double
verification.