Integration test the subscription api
This test is an end-to-end test of the Subscription API using the same
approach that we used for the integration test of Request. A
configurable Publish server is spun up and then we can make calls to the
client-facing API methods to make sure they work as the Readme says they
do.

The `test_delay` thing I'm doing here seems hacky, but it's the best way
I've figured out for making these tests work. The Publish server is
running in its own process and I can't control the startup time for that
process. Pausing after starting the server gives it time to spawn.
Pausing inside the tests ensures we get a message.

A more complex, but maybe less hacky approach would be to have an
inter-process socket that allows communication between the Publish
process and the Rspec process. The Publish process could use this socket
to tell Rspec when it's published something, and Rspec could use those
messages as flow-control for its tests. Something like 'once I get the
OK from Publisher, check the state of my objects.'

For this version of the tests I'm going to stick with the hacky
approach.